Add tunnel check and deactivate standby core sooner
The tunnel check pings some IP's on the other side of the tunnel. If one IP
replies the tunnel is considered up. Disable the tunnel check by setting
'tunnelCheckPerform' to false.

The standby core is restarted in non-producing mode when the block height is
reported again OR (instead AND) cncli reports 'ok'.