Fix bug where local core keeps running after deactivation
The use of parentheses in the for-loop causes a subshell to open which in turn
created an unwanted job-control issue in Bash where the 'sleep 60' command was
not executed.

Without the 60 second pause the local node did not have enough time to
reconfigure to non-producing mode before the keys were put back in place by the
next line of the script, leaving the node running in block producing mode.